refactor: Split business logic using hook in OrderPage
